fix(cart):fixed error of return of stock when deleting cart/Eze (#235)

+++ server/src/controllers/Cart/controllDeleteCartUser.js
@@ -1,13 +1,18 @@
-const { Carts } = require("../../db.js");
+const { Carts, Stocks } = require("../../db.js");
 
-const deleteCartUserController = async (id) => {
-  if (id) {
-    await Carts.destroy({
-      where: { UserId: id },
+const deleteCartUserController = async (id, sale) => {
+  if(sale === "false"){
+    const carts = await Carts.findAll({where:{UserId: id}});
+    carts.map(async(cart)=>{
+      const stock = await Stocks.findByPk(cart.StockId);
+      stock.amount += cart.quantity;
+      await stock.save();
     });
+  };
+  await Carts.destroy({where: { UserId: id }});
 
-    return { message: "CartUser deleted successfully." };
-  }
+  return { message: "CartUser deleted successfully." };
+  
 };
 
 module.exports = deleteCartUserController;
